[Refactor] Refactor imports
