+++ src/auth/auth.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './services/auth.service';
-import { AuthController } from './controllers';
-import { AtStrategy, LocalStrategy, RtStrategy } from './strategies';
 import { JwtModule } from '@nestjs/jwt';
+
+import { AtStrategy, LocalStrategy, RtStrategy } from './strategies';
 import { AtCookieConfig, RtCookieConfig } from '../configs';
+import { AuthService } from './services';
+import { AuthController } from './controllers';
 import { UserModule } from '../user/user.module';
 
 @Module({
