+++ src/auth/decorators/user-role.decorator.ts
@@ -1,4 +1,5 @@
 import { SetMetadata } from '@nestjs/common';
+
 import { Role } from '../../types';
 
 export const ROLES_KEY = 'roles';
