+++ src/auth/guards/role.guard.ts
@@ -2,8 +2,8 @@ import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 
 import { Role } from '../../types';
-import { IS_PUBLIC_KEY, ROLES_KEY } from '../decorators';
-import { UserService } from '../../user/services/user.service';
+import { ROLES_KEY } from '../decorators';
+import { UserService } from '../../user/services';
 
 @Injectable()
 export class RolesGuard implements CanActivate {
