+++ src/auth/services/auth.service.ts
@@ -1,16 +1,15 @@
 import { HttpStatus, Injectable, UnauthorizedException } from '@nestjs/common';
-import { AuthDto } from '../dto';
-import { UserEntity } from '../../user/entities/user.entity';
-import { hashData } from '../../utils';
-import * as argon2 from 'argon2';
 import { JwtService } from '@nestjs/jwt';
-import { CookieNames, Tokens } from '../../types';
-import { Response } from 'express';
-import { AtCookieConfig, RtCookieConfig } from '../../configs';
 import { ConfigService } from '@nestjs/config';
-import { UserService } from '../../user/services/user.service';
-import { UserResponse } from '../../types/user-response.type';
-import { AdminUserService } from 'src/user/services/admin-user.service';
+import * as argon2 from 'argon2';
+import { Response } from 'express';
+
+import { CookieNames, Tokens, UserResponse } from '../../types';
+import { AtCookieConfig, RtCookieConfig } from '../../configs';
+import { UserService, AdminUserService } from '../../user/services';
+import { AuthDto } from '../dto';
+import { UserEntity } from '../../user/entities';
+import { hashData } from '../../utils';
 
 @Injectable()
 export class AuthService {
