+++ src/auth/strategies/local.strategy.ts
@@ -2,8 +2,8 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
 
-import { AuthService } from '../services/auth.service';
-import { UserEntity } from '../../user/entities/user.entity';
+import { AuthService } from '../services';
+import { UserEntity } from '../../user/entities';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
