+++ src/auth/strategies/rt.strategy.ts
@@ -1,8 +1,9 @@
 import { PassportStrategy } from '@nestjs/passport';
-import { ExtractJwt, Strategy } from 'passport-jwt';
-import { Request } from 'express';
 import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
+import { ExtractJwt, Strategy } from 'passport-jwt';
+import { Request } from 'express';
+
 import { CookieNames } from '../../types';
 
 @Injectable()
