+++ src/database/database.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigService } from '@nestjs/config';
+
 import { getTypeOrmConfig } from '../configs';
 
 // TODO: Change TypeOrm to prizma for better return type safety
