+++ src/post/dto/update-post.dto.ts
@@ -1,4 +1,5 @@
 import { PartialType } from '@nestjs/mapped-types';
+
 import { CreatePostDto } from './create-post.dto';
 
 export class UpdatePostDto extends PartialType(CreatePostDto) {}
