+++ src/post/entities/post.entity.ts
@@ -7,7 +7,8 @@ import {
   PrimaryGeneratedColumn,
   UpdateDateColumn,
 } from 'typeorm';
-import { UserEntity } from '../../user/entities/user.entity';
+
+import { UserEntity } from '../../user/entities';
 
 @Entity({
   name: 'posts',
