+++ src/post/post.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { PostService } from './services/post.service';
-import { PostController, AdminPostController } from './controllers';
-import { UserModule } from '../user/user.module';
-import { PostEntity } from './entities';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AdminPostService } from './services/admin-post.service';
+
+import { PostService, AdminPostService } from './services';
+import { PostController, AdminPostController } from './controllers';
+import { PostEntity } from './entities';
+import { UserModule } from '../user/user.module';
 
 @Module({
   imports: [UserModule, TypeOrmModule.forFeature([PostEntity])],
