+++ src/types/user-response.type.ts
@@ -1,3 +1,3 @@
-import { UserEntity } from '../user/entities/user.entity';
+import { UserEntity } from '../user/entities';
 
 export type UserResponse = Pick<UserEntity, 'id' | 'email' | 'role'>;
