+++ src/user/controllers/index.ts
@@ -1,1 +1,2 @@
 export * from './user.controller';
+export * from './admin-user.controller';
