+++ src/user/services/index.ts
@@ -1,1 +1,3 @@
 export * from './user.service';
+export * from './admin-user.service';
+export * from './addRole.service';
