+++ src/user/user.module.ts
@@ -1,16 +1,14 @@
 import { Module } from '@nestjs/common';
-import { UserService } from './services/user.service';
-import { UserController } from './controllers';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserEntity } from './entities/user.entity';
-import { UserRoleEntity } from './entities/user-role.entity';
-import { AdminUserController } from './controllers/admin-user.controller';
-import { AdminUserService } from './services/admin-user.service';
+
+import { UserService, AdminUserService, AddRoleService } from './services';
+import { UserController, AdminUserController } from './controllers';
+import { UserEntity, UserRoleEntity } from './entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserEntity, UserRoleEntity])],
   controllers: [UserController, AdminUserController],
-  providers: [UserService, AdminUserService],
+  providers: [UserService, AdminUserService, AddRoleService],
   exports: [UserService, AdminUserService],
 })
 export class UserModule {}
